postModuleを追加し、postモジュールからdagger/hiltの依存を消す（ついでにImplクラス抹消）
